+++ lib.esm/abi/fragments.js
@@ -588,7 +588,7 @@ export class ParamType {
      *  Walks the **ParamType** with %%value%%, asynchronously calling
      *  %%process%% on each type, destructing the %%value%% recursively.
      *
-     *  This can be used to resolve ENS naes by walking and resolving each
+     *  This can be used to resolve ENS names by walking and resolving each
      *  ``"address"`` type.
      */
     async walkAsync(value, process) {